:lipstick: style(folder structure and ux,ui improvment): folder structure and ux,ui improvment
Hero and About Folder structure

Card component UI improvment

Button Hover action
