+++ apps/cms-landing-page/src/components/LandingPage/About/AboutHeader.tsx
@@ -2,7 +2,7 @@
 import { Rocket } from 'lucide-react';
 
 export const AboutHeader = () => (
-  <div className="max-w-3xl mx-auto text-center mb-16">
+  <div className="max-w-3xl mx-auto text-center mb-10">
     <div className="inline-flex justify-center mb-6">
       <div className="bg-gradient-to-r from-blue-600 to-purple-600 p-3 rounded-lg shadow-lg">
         <Rocket className="h-12 w-12 text-white" />
